Fix gwtc
