+++ src/peergos/server/tests/RetryStorageTests.java
@@ -136,7 +136,7 @@ public class RetryStorageTests {
         }
 
         @Override
-        public CompletableFuture<LinkRetrievalCounter.LinkCounts> getLinkCounts(String owner, LocalDateTime after, BatWithId mirrorBat) {
+        public CompletableFuture<LinkCounts> getLinkCounts(String owner, LocalDateTime after, BatWithId mirrorBat) {
             throw new IllegalStateException("Unimplemented!");
         }
