+++ src/peergos/shared/storage/ContentAddressedStorageProxy.java
@@ -1,6 +1,5 @@
 package peergos.shared.storage;
 
-import peergos.server.storage.*;
 import peergos.shared.cbor.*;
 import peergos.shared.crypto.*;
 import peergos.shared.crypto.hash.*;
@@ -29,7 +28,7 @@ public interface ContentAddressedStorageProxy {
 
     CompletableFuture<EncryptedCapability> getSecretLink(Multihash targetServerId, SecretLink link);
 
-    CompletableFuture<LinkRetrievalCounter.LinkCounts> getLinkCounts(Multihash targetServerId, String owner, LocalDateTime after, BatWithId mirrorBat);
+    CompletableFuture<LinkCounts> getLinkCounts(Multihash targetServerId, String owner, LocalDateTime after, BatWithId mirrorBat);
 
     CompletableFuture<List<Cid>> put(Multihash targetServerId,
                                      PublicKeyHash owner,
@@ -121,16 +120,16 @@ public interface ContentAddressedStorageProxy {
         }
 
         @Override
-        public CompletableFuture<LinkRetrievalCounter.LinkCounts> getLinkCounts(Multihash targetServerId,
-                                                                                String owner,
-                                                                                LocalDateTime after,
-                                                                                BatWithId mirrorBat) {
+        public CompletableFuture<LinkCounts> getLinkCounts(Multihash targetServerId,
+                                                           String owner,
+                                                           LocalDateTime after,
+                                                           BatWithId mirrorBat) {
             return poster.get(getProxyUrlPrefix(targetServerId) + apiPrefix
                     + "link/counts?after=" + after.toEpochSecond(ZoneOffset.UTC)
                     + "?bat=" + mirrorBat.encode()
                     + "&owner=" + owner)
                     .thenApply(CborObject::fromByteArray)
-                    .thenApply(LinkRetrievalCounter.LinkCounts::fromCbor);
+                    .thenApply(LinkCounts::fromCbor);
         }
 
         @Override
