+++ src/peergos/shared/user/UserSnapshot.java
@@ -1,13 +1,11 @@
 package peergos.shared.user;
 
-import peergos.server.storage.*;
 import peergos.shared.cbor.*;
 import peergos.shared.crypto.hash.*;
 import peergos.shared.social.*;
+import peergos.shared.storage.*;
 import peergos.shared.storage.auth.*;
-import peergos.shared.util.*;
 
-import java.time.*;
 import java.util.*;
 import java.util.stream.*;
 
@@ -17,13 +15,13 @@ public class UserSnapshot implements Cborable {
     public final List<BlindFollowRequest> pendingFollowReqs;
     public final List<BatWithId> mirrorBats;
     public final Optional<LoginData> login;
-    public final LinkRetrievalCounter.LinkCounts linkCounts;
+    public final LinkCounts linkCounts;
 
     public UserSnapshot(Map<PublicKeyHash, byte[]> pointerState,
                         List<BlindFollowRequest> pendingFollowReqs,
                         List<BatWithId> mirrorBats,
                         Optional<LoginData> login,
-                        LinkRetrievalCounter.LinkCounts linkCounts) {
+                        LinkCounts linkCounts) {
         this.pointerState = pointerState;
         this.pendingFollowReqs = pendingFollowReqs;
         this.mirrorBats = mirrorBats;
@@ -59,12 +57,12 @@ public class UserSnapshot implements Cborable {
                 .getMap(PublicKeyHash::fromCbor, c -> ((CborObject.CborByteArray)c).value);
         List<BatWithId> mirrorBats = m.getList("b", BatWithId::fromCbor);
         Optional<LoginData> login = m.getOptional("l", LoginData::fromCbor);
-        LinkRetrievalCounter.LinkCounts lc = m.get("lc", LinkRetrievalCounter.LinkCounts::fromCbor);
+        LinkCounts lc = m.get("lc", LinkCounts::fromCbor);
         return new UserSnapshot(pointerState, pendingFollowReqs, mirrorBats, login, lc);
     }
 
     public static UserSnapshot empty() {
         return new UserSnapshot(Collections.emptyMap(), Collections.emptyList(), Collections.emptyList(),
-                Optional.empty(), new LinkRetrievalCounter.LinkCounts(Collections.emptyMap()));
+                Optional.empty(), new LinkCounts(Collections.emptyMap()));
     }
 }
